+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Version History
 
+## Unreleased
+
+* [THORN-581] Update attribute cache from hardware read rather than software cache.
+
 ## 12.2.0
 
 * [SKB-1379] Move Discarded or Flagged Packet Count "<https://developer.skao.int/projects/ska-low-sps-tpm-api/en/3.2.0/generated/monitoring_points/dsp.html#discarded-or-flagged-packet-count>" to rollup into Spsstation health not MccsTile.